+++ src/Services/Calculator/Interest.Calculator.Application/Interfaces/IExecuteService.cs
@@ -5,6 +5,6 @@ namespace Interest.Calculator.Application.Interfaces
 {
     public interface IExecuteService
     {
-        Task<CalculationResponse> Execute(double initialValue, double months);
+        Task<CalculationResponse> Execute(double initialValue, int months);
     }
 }
